file_uploader_rest/pyproject.toml
[tool.poetry]
name = "django-rest-api"
version = "0.1.0"
description = ""
authors = ["Svetlov-al <s3t1qs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
celery = "^5.3.4"
django-redis = "^5.4.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
moviepy = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
